refactor(annotations): split geometry helpers
Move annotation bbox, point, quad, and numeric-array conversion helpers out of the annotation builder so field extraction and coordinate normalization remain easier to review independently.
